test: lmb: run lmb tests only manually
The LMB code has been changed so that the memory reservations and
allocations are now persistent and global. With this change, the
design of the LMB tests needs to be changed accordingly. For now, mark
the LMB tests to be run only manually. The tests won't be run as part
of the unit test suite, and thus would not interfere with the running
of the rest of the tests.

Signed-off-by: Sughosh Ganu <sughosh.ganu@linaro.org>