lmb: add nooverwrite versions of the lmb API
The LMB module is to be used as a backend for allocating and freeing
up memory requested from other modules like EFI. These memory requests
are different from the typical LMB reservations in that memory
required by the EFI module cannot be overwritten, or re-requested. Add
the nooverwrite versions of the LMB API functions for allocating and
freeing up memory. For now, these functions will be used by the EFI
memory module.

Signed-off-by: Sughosh Ganu <sughosh.ganu@linaro.org>