efi_memory: rename variable to highlight overlap with free memory
The functions to add a memory region to the EFI memory map have a
variable, overlap_only_ram, which is used to indicate if a new memory
region needs to be carved out from the free, or conventional
memory. This is a bit confusing, as other allocated memory regions,
like boot-services data and code are also part of the RAM
memory. Rename the variable to highlight the fact that it is the
free/conventional memory that is being referred to in this context.

Signed-off-by: Sughosh Ganu <sughosh.ganu@linaro.org>