common: memtop: add logic to detect ram_top
Add generic logic to determine the ram_top value for the
board. Earlier, this was achieved through LMB API's, but that has
since changed so that the LMB code is available only after
relocation. Replace those LMB calls with a single call to
get_mem_top() to determine the value of ram_top.

Signed-off-by: Sughosh Ganu <sughosh.ganu@linaro.org>